+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Added support for 1.20.2.
+- Fixed support for 1.16.5 and 1.18.2.
+- Added compatibility with Emotecraft.